refactor: MovieSeat 색상 지정방법 변경
