+++ app/src/main/java/woowacourse/movie/model/MovieSeats.kt
@@ -3,8 +3,8 @@ package woowacourse.movie.model
 class MovieSeats(private val rowSize: Int = 5, private val columnSize: Int = 4) {
     val seats =
         List(rowSize * columnSize) { index ->
-            val row = index / rowSize
-            val column = index % rowSize
+            val row = index / columnSize
+            val column = index % columnSize
             MovieSeat(row, column)
         }
 }
