+++ app/src/main/java/woowacourse/movie/seatselection/presenter/MovieSeatPresenter.kt
@@ -1,3 +1,5 @@
 package woowacourse.movie.seatselection.presenter
 
-class MovieSeatPresenter
+import woowacourse.movie.seatselection.presenter.contract.MovieSeatContract
+
+class MovieSeatPresenter(private val movieSeatContractView: MovieSeatContract.View) : MovieSeatContract.Presenter
