+++ app/src/main/res/values/colors.xml
@@ -7,5 +7,8 @@
     <color name="teal_700">#FF018786</color>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
-    <color name="yellow">#FAFF00</color>
+    <color name="selected">#FAFF00</color>
+    <color name="b_grade">#8E13EF</color>
+    <color name="s_grade">#19D358</color>
+    <color name="a_grade">#1B48E9</color>
 </resources>
